// src/shared/types.ts
export type EscopoAtuacao = 'INDIVIDUAL' | 'SETORIAL';

export interface Setor {
  id: string;
  nome: string;
  descricao?: string;
  ativo: boolean;
  // Metadata
  createdAt: string;
  updatedAt?: string;
}

export interface Cargo {
  id: string;
  nome: string;
  descricao?: string;
  setoresPermitidos: string[]; // IDs dos setores onde este cargo pode atuar
  escopo: EscopoAtuacao;
  ativo: boolean;
  // Metadata
  createdAt: string;
  updatedAt?: string;
}